Finish Settings Screen UI✅
- Create settings screen UI and settings_item.
